Change to the tanh function to behave like a sort of step function (see description)
Notice that:
C_1 + C_2 * 0.5 * (tanh(.) + 1 )
C_1 + C_2 * 0.5 * tanh(.) + 0.5 * C_2
C_3 + C_4 * tanh(.)
So as long as there is a constant in the equation we can adjust the tanh to be more interpretable. Even if there isn't we can just add a constant to make the  tanh behave like the step function.